requirements.txt
langchain
langchain-openai
langchain-pinecone
langchainhub
python-dotenv
black
langchain-community
pypdf
faiss-cpu